Rewrite timeline handling
This replaces the code in the single action evaluator with a general
class that calculates keyframes and progress values for interpolation.
This changes the semantics slightly: The final keyframe in an action
is _never_ reached; the action will loop back to the start on the frame
before the final keyframe. This may have consequences for keyframe
events, because an event on the final keyframe will never be evaluated.

Fix #17
Affects #7
Affects #9