Do not normalize weight vectors
The sum of weights in a weight vector should add up to 1.0. Previously,
the code was implemented such that it was normalizing vectors such
that the magnitude of the vector was 1.0. This is incorrect! The
weight values are independent and should not be treated as being in
any way related. This commit divides each component of the weight
vector by the sum of the components, yielding a vector with sum 1.0.