Initial example software mesh deformation
This implements some example code that deforms a mesh on the CPU.
The CaMeshDeformationMatrices has been rewritten for clarity at the
cost of allocating a few more matrices (allocated once and reused
forever). The CaSchemas type provides schema values for validating
meshes according to the as-yet-undocumented conventions used by
the jcalcium package.

Use smf 0.7.0.