CPU-side mesh deformation API
This provides a complete API and implementation for performing CPU-side
deformation of meshes, with an SMF-backed mesh implementation that
efficiently stores mesh data inside byte buffers (as opposed to using
large arrays of Java heap objects). The API also avoids retention by
design, allowing the programmer to completely control the lifetime
of mesh data.

An example program is included that demonstrates deformation of a mesh
in realtime.

Fix #21
Fix #15