Require that root joints have identity transforms
Root joints are now required to have an identity transform. That is,
they are required to have no rotation, a scale factor of 1.0, and a
translation of zero. This commit updates the Blender plugin to check
root bones, and also adds the same check in the skeleton compiler. It
also adds a couple of .blend test cases.

Fix #12
Affects #24